Moved DayViewHeader and getOverlapCount helper functions into utils/dateUtils.jsx
